fix: support migration for log entry formats
