+++ packages/log/src/clock.ts
@@ -24,6 +24,7 @@ SOFTWARE.
  */
 import {
 	field,
+	variant,
 	/* , variant */
 } from "@dao-xyz/borsh";
 import { hrtime } from "@peerbit/time";
@@ -58,7 +59,7 @@ function bigIntCoerce(input: any, fallback: any) {
 	return fallback;
 }
 
-/* @variant(0) */
+@variant(0)
 export class Timestamp {
 	@field({ type: "u64" })
 	wallTime: bigint;
@@ -215,7 +216,7 @@ export class ForwardJumpError extends Error {
 	}
 }
 
-/* @variant(0) */
+@variant(0)
 export class LamportClock {
 	@field({ type: Uint8Array })
 	id: Uint8Array;
