+++ packages/log/src/entry-type.ts
@@ -1,0 +1,4 @@
+export enum EntryType {
+	APPEND = 0, // Add more data
+	CUT = 1, // Delete or Create tombstone ... delete all nexts, i
+}
