+++ packages/log/src/index.ts
@@ -3,9 +3,14 @@ export * from "./log-sorting.js";
 export * from "./log-errors.js";
 export * from "./snapshot.js";
 export * from "./entry.js";
+export * from "./entry-type.js";
+export * from "./entry-with-refs.js";
+export * from "./entry-shallow.js";
 export * from "./utils.js";
 export * from "./clock.js";
 export * from "./encoding.js";
 export * from "./trim.js";
 export * from "./change.js";
+export * from "./entry-v0.js";
+export * from "./entry-create.js";
 export type { TrimToByteLengthOption, TrimToLengthOption } from "./trim.js";
