+++ packages/log/test/entry.spec.ts
@@ -9,7 +9,9 @@ import assert from "assert";
 import { expect } from "chai";
 import sodium from "libsodium-wrappers";
 import { LamportClock, Timestamp } from "../src/clock.js";
-import { Entry, Payload } from "../src/entry.js";
+import { createEntry } from "../src/entry-create.js";
+import { Entry } from "../src/entry.js";
+import { Payload } from "../src/payload.js";
 import { signKey } from "./fixtures/privateKey.js";
 import { JSON_ENCODING } from "./utils/encoding.js";
 
@@ -36,7 +38,7 @@ describe("entry", function () {
 			}
 		}
 		const create = () =>
-			Entry.create({
+			createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -62,7 +64,7 @@ describe("entry", function () {
 				timestamp: new Timestamp({ wallTime: 2n, logical: 3 }),
 			});
 
-			const entry = await Entry.create({
+			const entry = await createEntry({
 				store,
 				identity: signKey,
 
@@ -73,13 +75,13 @@ describe("entry", function () {
 				},
 			});
 			expect(entry.hash).to.equal(
-				"zb2rhZwBBmAvUNPkSjsKLSf1F41aSM4QP2XLQortvngCfhttg",
+				"zb2rhkp7iF9qm87YVdLERWfiChsfs8FhnTEqUB426kQtq3zro",
 			);
-			expect(entry.gid).equal(sha256Base64Sync(Buffer.from("a")));
+			expect(entry.meta.gid).equal(sha256Base64Sync(Buffer.from("a")));
 			expect(entry.meta.clock.equals(clock)).to.be.true;
 			expect(entry.payload.getValue()).to.deep.equal(new Uint8Array([1]));
-			expect(entry.next.length).equal(0);
-			expect(entry.size).equal(212);
+			expect(entry.meta.next.length).equal(0);
+			expect(entry.size).equal(215);
 		});
 
 		it("creates a entry with payload", async () => {
@@ -88,7 +90,7 @@ describe("entry", function () {
 				id: new Uint8Array([1, 2, 3]),
 				timestamp: new Timestamp({ wallTime: 2n, logical: 3 }),
 			});
-			const entry = await Entry.create({
+			const entry = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -100,19 +102,19 @@ describe("entry", function () {
 				encoding: JSON_ENCODING,
 			});
 			expect(entry.hash).to.equal(
-				"zb2rhkJGPvFC4NKfVACnJLgMN2wPPV8q2MFEaT4Buc89Sy9p8",
+				"zb2rhcD9YFepJetzdndygKFLjMorAQEEVremfwzxr2S8jkpwy",
 			);
 			expect(entry.payload.getValue()).to.deep.equal(payload);
-			expect(entry.gid).equal(sha256Base64Sync(Buffer.from("a")));
+			expect(entry.meta.gid).equal(sha256Base64Sync(Buffer.from("a")));
 			expect(entry.meta.clock.equals(clock)).to.be.true;
-			expect(entry.next.length).equal(0);
+			expect(entry.meta.next.length).equal(0);
 		});
 
 		it("creates a encrypted entry with payload", async () => {
 			const payload = new Uint8Array([1]);
 			const senderKey = await X25519Keypair.create();
 			const receiverKey = await X25519Keypair.create();
-			const entry = await Entry.create({
+			const entry = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -133,20 +135,20 @@ describe("entry", function () {
 			expect(entry.payload.getValue()).to.deep.equal(payload);
 
 			// We can not have a hash check because nonce of encryption will always change
-			expect(entry.gid).equal(sha256Base64Sync(Buffer.from("a")));
+			expect(entry.meta.gid).equal(sha256Base64Sync(Buffer.from("a")));
 			expect(entry.meta.clock.id).to.deep.equal(
 				new Ed25519PublicKey({
 					publicKey: signKey.publicKey.publicKey,
 				}).bytes,
 			);
 			expect(entry.meta.clock.timestamp.logical).equal(0);
-			expect(entry.next.length).equal(0);
+			expect(entry.meta.next.length).equal(0);
 		});
 
 		it("creates a entry with payload and next", async () => {
 			const payload1 = new Uint8Array([1]);
 			const payload2 = new Uint8Array([2]);
-			const entry1 = await Entry.create({
+			const entry1 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -159,7 +161,7 @@ describe("entry", function () {
 				},
 				data: payload1,
 			});
-			const entry2 = await Entry.create({
+			const entry2 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -173,14 +175,14 @@ describe("entry", function () {
 				data: payload2,
 			});
 			expect(entry2.payload.getValue()).to.deep.equal(payload2);
-			expect(entry2.next.length).equal(1);
+			expect(entry2.meta.next.length).equal(1);
 			expect(entry2.hash).to.equal(
-				"zb2rhhiDKq4NrW7LHgHuV5cRiP4J1ysJmboYSp9jjdd7MzLmu",
+				"zb2rhi1Jy97WwuumkKVYWb8ZdpjLiZVzHxCivcYZcs5Qn2N4T",
 			);
 		});
 
 		it("`next` parameter can be an array of strings", async () => {
-			const entry1 = await Entry.create({
+			const entry1 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -189,7 +191,7 @@ describe("entry", function () {
 				},
 				data: new Uint8Array([1]),
 			});
-			const entry2 = await Entry.create({
+			const entry2 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -198,11 +200,11 @@ describe("entry", function () {
 				},
 				data: new Uint8Array([2]),
 			});
-			assert.strictEqual(typeof entry2.next[0] === "string", true);
+			assert.strictEqual(typeof entry2.meta.next[0] === "string", true);
 		});
 
 		it("`next` parameter can be an array of Entry instances", async () => {
-			const entry1 = await Entry.create({
+			const entry1 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -211,7 +213,7 @@ describe("entry", function () {
 				},
 				data: new Uint8Array([1]),
 			});
-			const entry2 = await Entry.create({
+			const entry2 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -220,11 +222,11 @@ describe("entry", function () {
 				},
 				data: new Uint8Array([2]),
 			});
-			assert.strictEqual(typeof entry2.next[0] === "string", true);
+			assert.strictEqual(typeof entry2.meta.next[0] === "string", true);
 		});
 
 		it("can calculate join gid from `next` max chain length", async () => {
-			const entry0A = await Entry.create({
+			const entry0A = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -234,7 +236,7 @@ describe("entry", function () {
 				data: new Uint8Array([1]),
 			});
 
-			const entry1A = await Entry.create({
+			const entry1A = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -244,7 +246,7 @@ describe("entry", function () {
 				data: new Uint8Array([1]),
 			});
 
-			const entry1B = await Entry.create({
+			const entry1B = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -256,12 +258,12 @@ describe("entry", function () {
 				data: new Uint8Array([1]),
 			});
 
-			expect(entry1A.gid > entry1B.gid); // so that gid is not choosen because A has smaller gid
+			expect(entry1A.meta.gid > entry1B.meta.gid); // so that gid is not choosen because A has smaller gid
 			expect(entry1A.meta.clock.timestamp.logical).equal(
 				entry1B.meta.clock.timestamp.logical,
 			);
 
-			const entry2 = await Entry.create({
+			const entry2 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -270,13 +272,15 @@ describe("entry", function () {
 				},
 				data: new Uint8Array([2]),
 			});
-			expect(entry2.gid).equal(
-				entry1A.gid < entry1B.gid ? entry1A.gid : entry1B.gid,
+			expect(entry2.meta.gid).equal(
+				entry1A.meta.gid < entry1B.meta.gid
+					? entry1A.meta.gid
+					: entry1B.meta.gid,
 			);
 		});
 
 		it("can calculate join gid from `next` max clock", async () => {
-			const entry1A = await Entry.create({
+			const entry1A = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -286,7 +290,7 @@ describe("entry", function () {
 				data: new Uint8Array([1]),
 			});
 
-			const entry1B = await Entry.create({
+			const entry1B = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -298,12 +302,12 @@ describe("entry", function () {
 				data: new Uint8Array([1]),
 			});
 
-			expect(entry1B.gid > entry1A.gid); // so that gid is not choosen because B has smaller gid
+			expect(entry1B.meta.gid > entry1A.meta.gid); // so that gid is not choosen because B has smaller gid
 			expect(
 				entry1B.meta.clock.timestamp.compare(entry1A.meta.clock.timestamp),
 			).greaterThan(0);
 
-			const entry2 = await Entry.create({
+			const entry2 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -312,13 +316,15 @@ describe("entry", function () {
 				},
 				data: new Uint8Array([2]),
 			});
-			expect(entry2.gid).equal(
-				entry1A.gid < entry1B.gid ? entry1A.gid : entry1B.gid,
+			expect(entry2.meta.gid).equal(
+				entry1A.meta.gid < entry1B.meta.gid
+					? entry1A.meta.gid
+					: entry1B.meta.gid,
 			);
 		});
 
 		it("can calculate join gid from `next` gid comparison", async () => {
-			const entry1A = await Entry.create({
+			const entry1A = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -328,7 +334,7 @@ describe("entry", function () {
 				data: new Uint8Array([1]),
 			});
 
-			const entry1B = await Entry.create({
+			const entry1B = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -340,12 +346,12 @@ describe("entry", function () {
 				data: new Uint8Array([1]),
 			});
 
-			expect(entry1B.gid < entry1A.gid).to.be.true; // so that B is choosen because of gid
+			expect(entry1B.meta.gid < entry1A.meta.gid).to.be.true; // so that B is choosen because of gid
 			expect(entry1A.meta.clock.timestamp.logical).equal(
 				entry1B.meta.clock.timestamp.logical,
 			);
 
-			const entry2 = await Entry.create({
+			const entry2 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -354,13 +360,15 @@ describe("entry", function () {
 				},
 				data: new Uint8Array([2]),
 			});
-			expect(entry2.gid).equal(
-				entry1A.gid < entry1B.gid ? entry1A.gid : entry1B.gid,
+			expect(entry2.meta.gid).equal(
+				entry1A.meta.gid < entry1B.meta.gid
+					? entry1A.meta.gid
+					: entry1B.meta.gid,
 			);
 		});
 
 		it("can calculate reuse gid from `next`", async () => {
-			const entry1A = await Entry.create({
+			const entry1A = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -370,17 +378,17 @@ describe("entry", function () {
 				data: new Uint8Array([1]),
 			});
 
-			const entry1B = await Entry.create({
+			const entry1B = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
-					gid: entry1A.gid,
+					gid: entry1A.meta.gid,
 					next: [],
 				},
 				data: new Uint8Array([1]),
 			});
 
-			const entry2 = await Entry.create({
+			const entry2 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -389,12 +397,12 @@ describe("entry", function () {
 				},
 				data: new Uint8Array([2]),
 			});
-			expect(entry2.gid).equal(entry1A.gid);
-			expect(entry1A.gid).equal(entry1B.gid);
+			expect(entry2.meta.gid).equal(entry1A.meta.gid);
+			expect(entry1A.meta.gid).equal(entry1B.meta.gid);
 		});
 
 		it("will use next for gid instaed of gidSeed", async () => {
-			const entry1 = await Entry.create({
+			const entry1 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -404,7 +412,7 @@ describe("entry", function () {
 				data: new Uint8Array([1]),
 			});
 
-			const entry2 = await Entry.create({
+			const entry2 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -413,13 +421,13 @@ describe("entry", function () {
 				},
 				data: new Uint8Array([2]),
 			});
-			expect(entry2.gid).equal(entry1.gid);
+			expect(entry2.meta.gid).equal(entry1.meta.gid);
 		});
 
 		it("throws an error if data is not defined", async () => {
 			let err: any;
 			try {
-				await Entry.create({
+				await createEntry({
 					store,
 					identity: signKey,
 					meta: {
@@ -437,7 +445,7 @@ describe("entry", function () {
 		it("throws an error if next is not an array", async () => {
 			let err: any;
 			try {
-				await Entry.create({
+				await createEntry({
 					store,
 					identity: signKey,
 					meta: {
@@ -455,7 +463,7 @@ describe("entry", function () {
 
 	describe("toMultihash", () => {
 		it("returns an multihash", async () => {
-			const entry = await Entry.create({
+			const entry = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -473,7 +481,7 @@ describe("entry", function () {
 			const multihash = await Entry.toMultihash(store, entry);
 			expect(multihash).equal(hash);
 			expect(multihash).to.equal(
-				"zb2rhZwBBmAvUNPkSjsKLSf1F41aSM4QP2XLQortvngCfhttg",
+				"zb2rhkp7iF9qm87YVdLERWfiChsfs8FhnTEqUB426kQtq3zro",
 			);
 		});
 
@@ -482,7 +490,7 @@ describe("entry", function () {
 	it('throws an error if the object being passed is invalid', async () => {
 	  let err
 	  try {
-		const entry = await Entry.create({ store, identity: signKey, gidSeed:   'A', data: 'hello', next: [] })
+		const entry = await createEntry({ store, identity: signKey, gidSeed:   'A', data: 'hello', next: [] })
 		delete ((entry.metadata as MetadataSecure)._metadata as DecryptedThing<Metadata>)
 		await Entry.toMultihash(store, entry)
 	  } catch (e: any) {
@@ -496,7 +504,7 @@ describe("entry", function () {
 		it("creates a entry from hash", async () => {
 			const payload1 = new Uint8Array([1]);
 			const payload2 = new Uint8Array([2]);
-			const entry1 = await Entry.create({
+			const entry1 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -509,7 +517,7 @@ describe("entry", function () {
 				},
 				data: payload1,
 			});
-			const entry2 = await Entry.create({
+			const entry2 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -525,12 +533,12 @@ describe("entry", function () {
 			const final = await Entry.fromMultihash<Uint8Array>(store, entry2.hash);
 			final.init(entry2);
 			assert(final.equals(entry2));
-			expect(final.gid).equal(sha256Base64Sync(Buffer.from("a")));
+			expect(final.meta.gid).equal(sha256Base64Sync(Buffer.from("a")));
 			expect(final.payload.getValue()).to.deep.equal(payload2);
-			expect(final.next.length).equal(1);
-			expect(final.next[0]).equal(entry1.hash);
+			expect(final.meta.next.length).equal(1);
+			expect(final.meta.next[0]).equal(entry1.hash);
 			expect(final.hash).to.equal(
-				"zb2rhXNBbQprcTNTWmJCMW3ifnBjmtZ1DTS4LXHNP2VXKa18X",
+				"zb2rhcw32voNHstGRjjRE4X6Rb2oKS4tGtqzogERKvDKQf9iw",
 			);
 		});
 	});
@@ -539,7 +547,7 @@ describe("entry", function () {
 		it("returns true if entry has a child", async () => {
 			const payload1 = new Uint8Array([1]);
 			const payload2 = new Uint8Array([2]);
-			const entry1 = await Entry.create({
+			const entry1 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -548,7 +556,7 @@ describe("entry", function () {
 				},
 				data: payload1,
 			});
-			const entry2 = await Entry.create({
+			const entry2 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -563,7 +571,7 @@ describe("entry", function () {
 		it("returns false if entry does not have a child", async () => {
 			const payload1 = new Uint8Array([1]);
 			const payload2 = new Uint8Array([2]);
-			const entry1 = await Entry.create({
+			const entry1 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -572,7 +580,7 @@ describe("entry", function () {
 				},
 				data: payload1,
 			});
-			const entry2 = await Entry.create({
+			const entry2 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -581,7 +589,7 @@ describe("entry", function () {
 				},
 				data: payload2,
 			});
-			const entry3 = await Entry.create({
+			const entry3 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -599,7 +607,7 @@ describe("entry", function () {
 	describe("compare", () => {
 		it("returns true if entries are the same", async () => {
 			const payload1 = new Uint8Array([1]);
-			const entry1 = await Entry.create({
+			const entry1 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -612,7 +620,7 @@ describe("entry", function () {
 				},
 				data: payload1,
 			});
-			const entry2 = await Entry.create({
+			const entry2 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -631,7 +639,7 @@ describe("entry", function () {
 		it("returns true if entries are not the same", async () => {
 			const payload1 = new Uint8Array([0]);
 			const payload2 = new Uint8Array([1]);
-			const entry1 = await Entry.create({
+			const entry1 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -640,7 +648,7 @@ describe("entry", function () {
 				},
 				data: payload1,
 			});
-			const entry2 = await Entry.create({
+			const entry2 = await createEntry({
 				store,
 				identity: signKey,
 				meta: {
@@ -655,7 +663,7 @@ describe("entry", function () {
 
 	describe("verifySignatures", () => {
 		it("verifies", async () => {
-			const entry1 = await Entry.create({
+			const entry1 = await createEntry({
 				store,
 				identity: signKey,
 				data: new Uint8Array(0),
